crypto-news-buyer/requirements.txt
crewai
crewai-tools
duckduckgo-search
python-dotenv
